changes to blip 43 draft
A few changes:
1. Use consistent naming between peer_alt_addr and alt_addr.  I think
this should follow the trend of the bolts and spell out the full text
as well, i.e., alternate_address vs alt_addr.
2. Some MUST verbiage changed to SHOULD.  Unless the sending peer is
required to reject an incoming connection on the public address, this
needs to be a SHOULD.  (It's optional so we can't require that anyway.)
3. Use the same encoding as bolt7 for addresses rather than introducing
csv style here.  (Or otherwise document that, but I think it should stay
consistent.)
4. alt-bind-addr is CLN specific so doesn't need to go in the wire protocol
specification.